+++ tests/src/Unit/Resolver/MidnightResolverTest.php
@@ -15,6 +15,7 @@ use Drupal\Tests\UnitTestCase;
 use Drupal\timesup\Periodicity\MidnightResolver;
 use PHPUnit\Framework\Attributes\AllowMockObjectsWithoutExpectations;
 use PHPUnit\Framework\Attributes\DataProvider;
+use PHPUnit\Framework\Attributes\Group;
 
 /**
  * @coversDefaultClass \Drupal\timesup\Periodicity\MidnightResolver
@@ -22,6 +23,7 @@ use PHPUnit\Framework\Attributes\DataProvider;
  * @group timesup
  */
 #[AllowMockObjectsWithoutExpectations]
+#[Group('timesup')]
 final class MidnightResolverTest extends UnitTestCase {
   use InvokeMethodTrait;